requirements.txt
google-adk>=1.0.0
google-generativeai>=0.8.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
